refactor/#13: query 방식 -> prop 전달 방식으로 변경
