+++ components/register/SizeOption.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { OptionType } from 'pages/register/[...params]';
+import { OptionType } from 'pages/register';
 import styled, { css } from 'styled-components';
 import theme from 'styles/theme';
